--- src/types/teacher.ts
@@ -1,8 +1,0 @@
-export type teacherType = {
-  id: number | null;
-  name: string;
-  email: string;
-  birthDate: string;
-};
-
-export default teacherType;
